renamed popularMoviesSlice to moviesSlice
