+++ src/core/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "@redux-saga/core";
-import popularMoviesReducer from "../features/MovieList/popularMoviesSlice";
+import moviesReducer from "../features/MovieList/moviesSlice";
 import popularPeopleReducer from "../features/PersonList/popularPeopleSlice";
 import movieDetailsReducer from "../features/MovieDetails/movieDetailsSlice";
 import saga from "./saga";
@@ -9,7 +9,7 @@ const sagaMiddleware = createSagaMiddleware();
 
 const store = configureStore({
     reducer: {
-        popularMovies: popularMoviesReducer,
+        popularMovies: moviesReducer,
         popularPeople: popularPeopleReducer,
         movieDetails: movieDetailsReducer,
     },
