+++ src/features/MovieList/popularMoviesSaga.js
@@ -1,15 +1,15 @@
 import { takeEvery, debounce, call, put, all } from "@redux-saga/core/effects";
 
 import { getGenres, getPopularMoviesData, getSearchMoviesData } from "./popularMoviesAPI";
-import {  fetchMoviesError, fetchMoviesSuccess, fetchPopularMoviesLoading, fetchSearchMoviesLoading } from "./popularMoviesSlice";
+import {  fetchMoviesError, fetchMoviesSuccess, fetchPopularMoviesLoading, fetchSearchMoviesLoading } from "./moviesSlice";
 
 function* fetchPopularMoviesDataHandler({ payload: page }) {
     try {
-        const [moviesData, popularMoviesGenres] = yield all([
+        const [moviesData, moviesGenres] = yield all([
             call(getPopularMoviesData, page),
             call(getGenres),
         ]);
-        yield put(fetchMoviesSuccess({ moviesData, popularMoviesGenres }));
+        yield put(fetchMoviesSuccess({ moviesData, moviesGenres }));
     } catch (error) {
         yield put(fetchMoviesError());
     }
@@ -17,11 +17,11 @@ function* fetchPopularMoviesDataHandler({ payload: page }) {
 
 function* fetchSearchMoviesDataHandler({ payload: location }) {
     try {
-        const [moviesData, popularMoviesGenres] = yield all([
+        const [moviesData, moviesGenres] = yield all([
             call(getSearchMoviesData, location),
             call(getGenres),
         ]);
-        yield put(fetchMoviesSuccess({ moviesData, popularMoviesGenres }));
+        yield put(fetchMoviesSuccess({ moviesData, moviesGenres }));
     } catch (error) {
         yield put(fetchMoviesError());
     }
